Relocation hpa package
