+++ bcs-app/backend/kube_core/hpa/constants.py
