+++ bcs-app/backend/kube_core/hpa/utils.py
